+++ lib/_core/precached_icons.dart
@@ -14,10 +14,15 @@ class PrecachedIcons {
   static const navBarMicOffLineIcon = 'assets/icons/nav_bar/line_mic_off.svg';
   static const navBarPhoneIcon = 'assets/icons/nav_bar/phone.svg';
   static const navBarPlanetIcon = 'assets/icons/nav_bar/planet.svg';
+  static const deviceMenuArrow = 'assets/icons/menu/Caret_Down_MD.svg';
+  static const inputOpenedEyeIcon = 'assets/icons/input/opened_eye.svg';
+  static const inputClosedEyeIcon = 'assets/icons/input/closed_eye.svg';
   static const micButton = 'assets/icons/mic/microphone_button.svg';
   static const settingsButton = 'assets/icons/circle_button/settings.svg';
   static const leaveCallButton = 'assets/icons/circle_button/leave_call.svg';
   static const startCallButton = 'assets/icons/circle_button/start_call.svg';
+  static const participantMicIcon =
+      'assets/icons/participants/mic_muted_mini.svg';
 
   List<SvgAssetLoader> get iconsToPrecache => const [
         SvgAssetLoader(copyDefaultIcon),
@@ -29,10 +34,14 @@ class PrecachedIcons {
         SvgAssetLoader(navBarMicOffLineIcon),
         SvgAssetLoader(navBarPhoneIcon),
         SvgAssetLoader(navBarPlanetIcon),
+        SvgAssetLoader(deviceMenuArrow),
+        SvgAssetLoader(inputOpenedEyeIcon),
+        SvgAssetLoader(inputClosedEyeIcon),
         SvgAssetLoader(micButton),
         SvgAssetLoader(settingsButton),
         SvgAssetLoader(leaveCallButton),
         SvgAssetLoader(startCallButton),
+        SvgAssetLoader(participantMicIcon),
       ];
 
   void precache(BuildContext context) {