+++ lib/_stories/widgets/wave_flower_loader_story.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart';
+import 'package:md_ui_kit/widgets/wave_flower_loader.dart';
+
+class WaveFlowerLoaderStory extends StatelessWidget {
+  const WaveFlowerLoaderStory({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return const WaveFlowerLoader();
+  }
+}